src/converter-svc/requirements.txt
pika
pymongo
moviepy
python-dotenv